Fixed weird transmitter issue
